+++ src/page-layout-tabs/document-template-editor.page-layout-tab.ts
@@ -17,7 +17,7 @@ export default definePageLayoutTab({
   widgets: [
     {
       universalIdentifier: DOCUMENT_TEMPLATE_PAGE_LAYOUT_EDITOR_WIDGET_UNIVERSAL_IDENTIFIER,
-      title: 'Template Editor',
+      title: 'Template Preview',
       type: 'FRONT_COMPONENT',
       gridPosition: { row: 0, column: 0, rowSpan: 4, columnSpan: 4 },
       configuration: {